requirements.txt
annotated-types==0.7.0
anyio==4.7.0
dnspython==2.7.0
fastapi==0.115.6
fastapi-utils==0.8.0
idna==3.10
psutil==5.9.8
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.10.1
python-dotenv==1.0.1
redis==5.2.1
slack_sdk==3.34.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
